+++ samples/sample_c/module_sample/utils/util_file.c
@@ -89,8 +89,8 @@ T_DjiReturnCode UtilFile_GetFileSizeByPath(const char *filePath, uint32_t *fileS
     return DJI_ERROR_SYSTEM_MODULE_CODE_SUCCESS;
 }
 
-T_DjiReturnCode UtilFile_GetFileDataByPath(const char *filePath, uint32_t offset, uint16_t len,
-                                           uint8_t *data, uint16_t *realLen)
+T_DjiReturnCode UtilFile_GetFileDataByPath(const char *filePath, uint32_t offset, uint32_t len,
+                                           uint8_t *data, uint32_t *realLen)
 {
     FILE *pF;
     T_DjiReturnCode psdkStat;
